Moved parsers to Dialogs namespace
